add fixes for e1.6.0
1) fix compilation errors: 
- ItemObject.All -> Items.All
- Use ClanChangedKingdom instead of deleted MercenaryClanChangedKingdom in ForceNoKingdomBehaviour

2) Major Talewords refactor for clans and mobile parties
- Clans are no longer registered in MBObjectManager but in the CampaignManager. 
Clans are now loaded later on in the process that's why it has been added in the "OnGameInitializationFinished" call. 
Moreover,  a "parseFaction" method was created in order to read the faction name from the xml instead of using the MBObjectManager read xml method.
- Two instances of SpawnDataManager were used which caused issues related to the loading of clans. Now, every object uses the instance from the Main class. 

3) Party Speed model

The current model used for PartySpeed in CustomSpawns overrides Talewords implementation and is quite outdated (new character and faction perks).
Copy pasting the latest implementation caused some new issues.
It's been faster to use a harmony patch instead. Therefore, there is no need to update the code after a new update and increases mod compatibility (known mod issue).
Although, the patch has a slightly different implementation as the former one. The patch will lock the final speed between any defined minimum and maximum speed. 
Moreover, I could not really understand the base speed override real purpose. Maybe, the way Talewords speed was implemented changed and the mod's impl. has not been updated accordingly, It would make sense if the base speed was the actual party speed on the map. 
That's why the patch overrides the given value which will still be affected by terrain and other factors (night...).

4) Issue with e1.6.0
Creating a new MobileParty with no PartyComponent (non bandit parties) causes the newly spawned parties to have an empty tracker on the map (trackers are the same as the player's clan war parties and caravans)

A harmony patch RemovePartyTrackersFromNonBanditPartiesPatch resolves the annoying issue.